requirements.txt
gunicorn
flask
pillow
google-cloud-storage
google-cloud-firestore